update branch calculation

+++ fuzzer/src/feedback.rs
@@ -16,14 +16,15 @@ pub struct Branch(usize);
 
 impl From<(Block, Block)> for Branch {
     fn from((b1, b2): (Block, Block)) -> Self {
-        let mut a = b1.0;
+        let mut a = b1.0 as u32;
         // hash algorithm from syzkaller
         a = (a ^ 61) ^ (a >> 16);
         a = a + (a << 3);
         a = a ^ (a >> 4);
         a *= 0x27d4_eb2d;
         a = a ^ (a >> 15);
-        Self(a ^ b2.0)
+
+        Self(a as usize ^ b2.0)
     }
 }
